Implements verification code for user password reset. Must use Sarah's phone number in account creation
