+++ app/models/code_generator.rb
@@ -1,4 +1,4 @@
-module CodeGenerator
+class CodeGenerator
   def self.generate
     rand(100000...999999).to_s
   end
