+++ app/views/trips/index.html.erb
@@ -10,7 +10,7 @@
 <%= link_to "Edit Email", edit_user_path(current_user) %>
 <%= link_to "Edit Phone", edit_user_path(current_user) %>
 <%= link_to "Update Image", edit_user_path(current_user) %>
-<%= link_to "Change Password", users_password_reset_path(current_user.username), method: :get %>
+<%= link_to "Change Password", users_password_reset_path(current_user.id), method: :get %>
 #EDIT WHEN WE CREATE THIS PATH
 <%= "View My Reviews" %>
 <%= link_to "Delete Account", user_path(current_user), method: :delete %>
